Dataset 2019/07/02
This is a dataset used in the analysis. It was created using code from [python_data.py](https://github.com/Pjarzabek/MoviesRuntimes/blob/master/movies_data.py). It was scraped from IMDb database on July 2nd, 2019. If you are interested in replicating my findings on your own, use this dataset.